@@ -1,6 +1,6 @@
-module github.com/glimpsovstar/packer-provisioner-ansible-aap
+module github.com/rptcloud/packer-provisioner-ansible-aap
 
-go 1.23.0
+go 1.24.0
 
 toolchain go1.24.2
 
